+++ config/test.exs
@@ -12,4 +12,4 @@ config :ex_aws, :hackney_opts,
   follow_redirect: true,
   recv_timeout: 30_000
 
-config :ex_aws, json_codec: Jason
+config :ex_aws, json_codec: JSON